swh.indexer.language: Improve edge cases policy
Before this commit, the following situation could arise:
- no encoding detected (we cannot do much after that yet)
- no result after multiple tryouts to compute the lang

This resulted in exceptions at runtime.

So, for now, we check those situations and we reference the sha1 with
lang None. The purpose is to permit an ulterior scheduling of those
later on (e.g. when we improve the encoding policy reuse for example).