indexer: Change from policy_update to a boolean named skip_existing
`policy_update` is a variable to determine whether to update or ignore
duplicates. Prior to this commit, this was conflated with the
incremental nature of the indexer.  This declares a meaningful name
instead 'incremental' and force the db policy update writing to always
update conflicting data (if any).

This `policy_update` variable should probably go away.