proper pagination for IndexerStorage.origin_intrinsic_metadata_search_by_producer
replace the 'start' argument by a 'page_token' which is now expected to be
an (not so) opaque token used for pagination.

The 'end' argument is dropped.

The method also returns a dict now (instead of generating the origins) with
2 keys: `next_page_token` and `origins`.

Adapt tests and the list_origins_by_producer() helper function (in cli.py)
accordingly.