BaseRow.unique_key: Don't crash when indexer_configuration_id is None.
This always happens when writing to Kafka, as the Kafka writer sets it
to None at the same it it injects the 'tool' data.

This was not caught by tests because they use the in-mem writer;
which did not call unique_key() at all in swh-journal<=v0.5.1
(but future versions will).