Fix crash when RawExtrinsicMetadata target new origins
RawExtrinsicMetadata contain a swh:1:ori: identifier of the origin,
which the indexer needs to resolve, by querying its storage replica.

Because RawExtrinsicMetadata are created by loaders, they are often
created shortly after the origin is created by the corresponding lister,
so the origin may not be known to the storage replica used by the
indexer, causing this function to crash.

Waiting 10s seems to be good enough when run on my computer with
production data and moma's replica; so I set it to 60s just to be safe.