tasks: Fix registration in scheduler database
Docstring is mandatory for a celery task function as its value
is inserted in the required description column of the task_type
table in the scheduler database.

Celery task function name is also used as task name (with underscores
replaced by dashes) so ensure task function names match the task types
registered in production scheduler database.