indexer: Use CompositeObjId as object id type for ContentIndexer
Previously, content indexers were only using content sha1 as object
identifier but this could lead to errors in production when attempting
to fetch content data from an object storage where key is the sha256
checksum of the content (winery for instance).

So ensure to use a CompositeObjId dictionary as content object identifier
to avoid these kind of issues when indexing content objects.