bibtex: Add option to force CodeMeta v3.0 context in codemeta_to_bibtex
Citation is failing to be generated if the JSON-LD context is missing
or malformed or when an unknown context URL is involved.

This can now be mitigated by setting the new force_codemeta_context
parameter to True as using CodeMeta v3.0 context is sufficient to
extract citation data.

Related to #4748 and #4750.